I mistakenly set the $_POST['author'] and $_POST['email'] variables for every request, regardless of whether I was actually doing OpenID login. This may have broke stuff in the core somewhere.
More importantly, this caused all anononymous comments to be stamped:
Name: openid
Email: placeholder@example.com

This is terrible from a usability and spam fighting perspective.

Changed hook method to filter get_option('require_name_email'), and return false if $_POST['openid_url'] is set. If this is the case, we're going to either start an OpenID redirect or dump to the login screen. Hole is closed.