less mucking around with the PHP include path
the calls to restore_include_path seem to be causing problems with other
WordPress plugins (like [instapress][]) because it resets the include
path to the master value set in php.ini, thereby losing any
modifications made by other plugins.  So instead of constantly setting
the include path and then restoring it, we set it once at the beginning
of loading the plugin, and then just leave it alone.  In order to
prevent any accidental conflicts, I'm also moving the OpenID library
into a dedicated lib directory to ensure that we're only adding the
library to the include path, not all files in the plugin directory.

[instapress]: http://wordpress.org/support/topic/plugin-openid-openid-conflicting-with-instapress-plugin